Create Trends in Startups
It’s my first day as a TechCrunch reporter.My first task is to write an article on the rising trends in the startup world.

To get me started with my research, my boss emailed me a project.sqlite file that contains a table called startups. It is a portfolio of some of the biggest names in the industry.

I write queries with aggregate functions to retrieve some interesting insights about these companies.